add macOS support in install script

+++ install.sh
@@ -6,7 +6,7 @@ mkdir -p ~/.bin
 mkdir -p ~/.ssh
 
 # ssh
-echo "AddKeysToAgent yes" >> ~/.ssh/config
+[ -z "$(grep "AddKeysToAgent yes" ~/.ssh/config || "")" ] && echo "AddKeysToAgent yes" >> ~/.ssh/config
 chmod 755 ~/.ssh/config
 
 # diff-so-fancy
@@ -16,7 +16,11 @@ chmod +x ~/.bin/diff-so-fancy
 ##########
 # git
 ##########
-sudo apt-get install -y git
+if [ $(uname -s) = "Darwin" ]; then
+    brew install git
+else
+    sudo apt-get install -y git
+fi
 git config --global alias.can 'commit --amend --no-edit'
 git config --global alias.pf 'push --force-with-lease'
 
@@ -41,7 +45,11 @@ git config --global color.diff.whitespace 'red reverse'
 ##########
 
 # Install zsh (Linux only)
-sudo apt-get install -y zsh
+if [ $(uname -s) = "Darwin" ]; then
+    brew install zsh
+else
+    sudo apt-get install -y zsh
+fi
 
 # Antibody
 curl -sL git.io/antibody | sh -s
@@ -53,9 +61,6 @@ touch ~/.zsh/custom/additional.zsh
 # Symlink zshrc
 ln -sfn $(pwd)/zsh/.zshrc ~/.zshrc
 
-# Create static antibody bundle
-antibody bundle < $(pwd)/zsh/plugins.txt > ~/.zsh-plugins.sh
-
 # Load zsh
 chsh -s $(which zsh)
 
@@ -66,9 +71,13 @@ chsh -s $(which zsh)
 # Install if not already present
 command -v node > /dev/null 2>&1
 if [ $? -ne 0 ]; then
-    echo 'Node not found. Installing 8.x...'
-    curl -sL https://deb.nodesource.com/setup_8.x | sudo -E bash -
-    sudo apt install -y nodejs
+    echo 'Node not found. Installing...'
+    if [ $(uname -s) = "Darwin" ]; then
+        brew install node
+    else
+        curl -sL https://deb.nodesource.com/setup_8.x | sudo -E bash -
+        sudo apt install -y nodejs
+    fi
 else
     echo 'Node is already installed'
 fi
@@ -81,13 +90,18 @@ sudo n lts
 # Neovim
 ##########
 
-# Install (Debian-based only)
-sudo add-apt-repository ppa:neovim-ppa/stable
-sudo apt update
-sudo apt upgrade -y
-sudo apt install -y neovim \
-    python-dev python-pip \
-    python3-dev python3-pip
+# Install
+if [ $(uname -s) = "Darwin" ]; then
+    brew install neovim
+    brew install python
+else
+    sudo add-apt-repository ppa:neovim-ppa/stable
+    sudo apt update
+    sudo apt upgrade -y
+    sudo apt install -y neovim \
+        python-dev python-pip \
+        python3-dev python3-pip
+fi
 pip3 install neovim
 pip install neovim
 
@@ -98,12 +112,12 @@ sudo npm i -g typescript # Nvim-typescript
 sudo npm i -g typescript-formatter # Neoformat
 sudo npm i -g tern # Javascript support
 
+# link to the right place
+ln -sfn $(pwd)/nvim ~/.config/nvim
+
 # Set up
 nvim -c 'PlugInstall | PlugUpdate'
 nvim -c 'UpdateRemotePlugins'
-
-# link to the right place
-ln -sfn $(pwd)/nvim ~/.config/nvim
 
 # Use as the default editor for git
 git config --global core.editor nvim
@@ -116,22 +130,26 @@ git config --global core.editor nvim
 if [[ $(tmux -V) = *2.6 ]]; then
     echo 'Tmux is already at the correct version'
 else
-    command -v tmux > /dev/null 2>&1
-    if [ $? -eq 0 ]; then
-        echo 'Found an existing installation of tmux. Removing it...'
+    if [ $(uname -s) = "Darwin" ]; then
+        brew install tmux
+    else
+        command -v tmux > /dev/null 2>&1
+        if [ $? -eq 0 ]; then
+            echo 'Found an existing installation of tmux. Removing it...'
 
-        # Debian-based only
-        sudo apt remove tmux
+            # Debian-based only
+            sudo apt remove tmux
+        fi
+
+        echo 'Installing tmux'
+        sudo apt-get install -y libevent-dev libncurses-dev build-essential
+        wget https://github.com/tmux/tmux/releases/download/2.7/tmux-2.7.tar.gz
+        tar -xzvf tmux-2.7.tar.gz
+        cd tmux-2.7
+        ./configure && make
+        sudo make install
+        cd -
     fi
-
-    echo 'Installing tmux'
-    sudo apt-get install -y libevent-dev libncurses-dev build-essential
-    wget https://github.com/tmux/tmux/releases/download/2.7/tmux-2.7.tar.gz
-    tar -xzvf tmux-2.7.tar.gz
-    cd tmux-2.7
-    ./configure && make
-    sudo make install
-    cd -
 
     # Alias for TrueColor support
     alias tmux="env TERM=xterm-256color tmux"
